Fix: 한 번에 하나의 video player만 메모리에 올라가도록 개선 (#243)
