+++ src/features/tutorial/components/TutorialListItem.tsx
@@ -1,10 +1,9 @@
 import { COLOR } from '@/src/constants/theme';
 import { Body2, MiddleTitle } from '@/src/core/Txt';
-import { VideoView, useVideoPlayer } from 'expo-video';
-import { useEffect } from 'react';
 import type { CarouselRenderItemInfo } from 'react-native-reanimated-carousel/lib/typescript/types';
 import styled from 'styled-components/native';
 import type { TUTORIAL_LIST_DATA } from '../constants';
+import TutorialVideoPlayer from './TutorialVideoPlayer';
 
 type Props = CarouselRenderItemInfo<(typeof TUTORIAL_LIST_DATA)[number]> & {
   isActive: boolean;
@@ -13,23 +12,9 @@ type Props = CarouselRenderItemInfo<(typeof TUTORIAL_LIST_DATA)[number]> & {
 const TutorialListItem = ({ item, isActive }: Props) => {
   const { title, description, video } = item;
 
-  const videoPlayer = useVideoPlayer(video, (player) => {
-    player.loop = true;
-    player.muted = true;
-  });
-
-  useEffect(() => {
-    if (isActive) {
-      videoPlayer.play();
-    } else {
-      videoPlayer.pause();
-      videoPlayer.currentTime = 0;
-    }
-  }, [videoPlayer, isActive]);
-
   return (
     <Slide>
-      <SlideVideo player={videoPlayer} contentFit="contain" nativeControls={false} />
+      {isActive ? <TutorialVideoPlayer source={video} /> : <TutorialVideoPlaceholder />}
       <TextContainer>
         <StyledMiddleTitle color="title">{title}</StyledMiddleTitle>
         <StyledBody2 color="sub1">{description}</StyledBody2>
@@ -47,11 +32,10 @@ const Slide = styled.View`
   background-color: ${COLOR.background};
 `;
 
-const SlideVideo = styled(VideoView)`
+const TutorialVideoPlaceholder = styled.View`
   flex: 1;
   width: 100%;
   height: 90%;
-  background-color: ${COLOR.background};
 `;
 
 const TextContainer = styled.View`
